# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Swagger documentation.",
        "title": "Analysis Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/analyze": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyze"
                ],
                "summary": "Create new analyze",
                "parameters": [
                    {
                        "description": "Analyze creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_hse-revizor_analysis-service_internal_pkg_router_dto.CreateAnalyzeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_hse-revizor_analysis-service_internal_pkg_router_dto.AnalyzeResponse"
                        }
                    }
                }
            }
        },
        "/api/analyze/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyze"
                ],
                "summary": "Get analyze by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Analyze ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_hse-revizor_analysis-service_internal_pkg_router_dto.AnalyzeResponse"
                        }
                    }
                }
            }
        },
        "/api/projects/{project_id}/analyzes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyze"
                ],
                "summary": "Get analyzes by project ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_hse-revizor_analysis-service_internal_pkg_router_dto.AnalyzeResponse"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_hse-revizor_analysis-service_internal_pkg_router_dto.AnalyzeResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "rule_results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hse-revizor_analysis-service_internal_pkg_router_dto.AnalyzeResultResponse"
                    }
                },
                "ruleset_id": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_hse-revizor_analysis-service_internal_pkg_router_dto.AnalyzeResultResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "rule_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_hse-revizor_analysis-service_internal_pkg_router_dto.CreateAnalyzeRequest": {
            "type": "object",
            "required": [
                "project_id",
                "ruleset_id"
            ],
            "properties": {
                "project_id": {
                    "type": "string"
                },
                "ruleset_id": {
                    "type": "string"
                }
            }
        }
    }
}